Add playbooks to do independent stops and starts of hypervisors
This adds stop and start playbooks (basically refactors of the restart
playbook) that can be run to bring down a hypervisor and then to start
it back up and restart the instances that were running on it. The use
case here is to allow HVs to be powered off for firmware updates while
still knowing what instances should be relaunched on them when they are
powered back up.