Hardcode certain service ports into the firewall configuration
iptables supports a maximum of 15 ports in a multiport specification.
In order to reduce the length of the list, certain ports are now
hardcoded into another multiport statement for services that we know we
will always be running (Apache on HTTP/HTTPS, public Keystone, Nova,
Cinder, Glance).

Keystone admin port 35357 is not included in this list in case we want
to easily remove it later on without requiring code changes.