Move "option nolinger" to only affect MySQL
"option nolinger" was originally added as part of a set of changes
for stabilizing MySQL access through HAProxy. However, the presence of
this option appears to be responsible for triggering TCP RSTs when
accessing the OpenStack APIs. "option nolinger" has been moved from the
defaults into the mysql-galera listener configuration, so that it will
continue to apply to MySQL while not impacting the OpenStack APIs.