Add instance information to DNS TXT records
This extends the DNS backfiller to add TXT records for qualifying
instances that contain the instance UUID and information on the
instance's Keystone project, so that a responsible party for an
instance can be easily located without involving cloud operators.

Querying the information is as easy as using dig to get the TXT record
associated with the instance's A record (if you have the CNAME, you must
first resolve that to the A record being pointed to).

This changeset also fixes a few places in the DNS backfiller where
error-handling code had mistakes in it and would fail if invoked.