Explicitly define the order of AZ joins/departs if in maintenance
I tried to be too cute with how entering/exiting maintenance would work.
In practice, it didn't actually work because OpenStack would still blow
up if you tried to put a node into two AZs at the same time. This gets
rid of that and explicitly defines the order in which AZs should be
joined and departed when entering and exiting maintenance.