Adjust recipes requiring search to load node lists lazily
In order to fix a problem where non-monitor nodes were being added to
the mon list in ceph.conf, the search functions were modified to not
unnecessarily push the current node onto the list. This resulted in a
number of issues in areas where we apparently relied on the broken
behavior, necessitating additional fixes in other recipes to wrap
template variable sections in `lazy {}`, moving some bash resources to
ruby_block resources, and generally ensuring that anything that deals
with node searches is processed at runtime and not compile time.